Moved getting contacts to main and added navigation for selecting contacts for a meeting (not finished yet)
