+++ lib/main.dart
@@ -2,13 +2,16 @@ import 'package:flutter/material.dart';
 import 'package:google_maps_flutter/google_maps_flutter.dart';
 import 'package:google_maps_flutter_platform_interface/google_maps_flutter_platform_interface.dart';
 import 'package:google_maps_webservice/places.dart';
+import 'package:lets_meet/src/model/contact.dart';
 import 'package:lets_meet/src/screens/edit_meeting_screen.dart';
-import 'package:lets_meet/src/widgets/contacts_main.dart';
+import 'package:lets_meet/src/screens/contacts_screen.dart';
 import 'package:maps_toolkit/maps_toolkit.dart' as mt;
 import 'src/screens/organize_meeting_screen.dart';
 import 'src/api/api_current_meetings.dart';
 import 'dart:async';
 import 'src/api/api_favourite_places.dart';
+import 'package:permission_handler/permission_handler.dart';
+import 'src/api/api_contacts.dart';
 
 // Uncomment lines 3 and 6 to view the visual layout at runtime.
 //import 'package:flutter/rendering.dart' show debugPaintSizeEnabled;
@@ -46,6 +49,23 @@ class _MapScreenState extends State<MapScreen> {
   LatLng _tappedLocation = const LatLng(0,0);
   Color mainColor = Colors.orange.shade700;
   Color secondaryColor = Colors.orange.shade100;
+  final Permission _permission = Permission.contacts;
+  final ContactsApi _contactsApi = ContactsApi();
+  List<Contact> _contactsList = [];
+
+  @override
+  void initState() {
+    super.initState();
+    initializeContacts();
+  }
+
+  Future<void> requestPermission() async {
+    final status = await _permission.request();
+
+    setState(() {
+
+    });
+  }
 
   void _onMapCreated(GoogleMapController controller) {
     mapController = controller;
@@ -404,7 +424,7 @@ class _MapScreenState extends State<MapScreen> {
                         context,
                         MaterialPageRoute(
                           builder: (context) =>
-                              OrganizeMeeting(details),
+                              OrganizeMeeting(place: details, contactsList: _contactsList),
                         ),
                       ),
                     },
@@ -499,6 +519,27 @@ class _MapScreenState extends State<MapScreen> {
     }
     setState(() {
       _fitMapToMarkers();
+    });
+  }
+
+  Future<void> initializeContacts() async {
+    _contactsList = await _contactsApi.getContactsLocal();
+    setState(() {
+
+    });
+  }
+
+  Future<void> getPhoneContacts() async {
+    //_csContacts = await cs.ContactsService.getContacts(withThumbnails: true);
+
+    setState(() {
+
+    });
+  }
+
+  void updateContactsList(List<Contact> updatedList) {
+    setState(() {
+      _contactsList = updatedList;
     });
   }
 
@@ -689,7 +730,7 @@ class _MapScreenState extends State<MapScreen> {
                       context,
                       MaterialPageRoute(
                         builder: (context) =>
-                            const ContactsManagement(),
+                            ContactsManagement(contactList: _contactsList, updateContactsList: updateContactsList),
                       ),
                     );
                   },
